requirements.txt
-i https://pypi.org/simple
anyio==4.0.0; python_version >= '3.8'
appdirs==1.4.4
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
arrow==1.3.0; python_version >= '3.8'
asttokens==2.4.0
async-lru==2.0.4; python_version >= '3.8'
attrs==23.1.0; python_version >= '3.7'
babel==2.13.0; python_version >= '3.7'
backcall==0.2.0
beautifulsoup4==4.12.2; python_version >= '3.6'
bleach==6.1.0; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.0; python_version >= '3.7'
comm==0.1.4; python_version >= '3.6'
configargparse==1.7; python_version >= '3.5'
connection-pool==0.0.3
contourpy==1.0.7
cycler==0.11.0
datrie==0.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
debugpy==1.8.0; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
docutils==0.20.1; python_version >= '3.7'
dpath==2.1.6; python_version >= '3.7'
exceptiongroup==1.1.3; python_version < '3.11'
executing==2.0.0
fastjsonschema==2.18.1
fonttools==4.40.0
fqdn==1.5.1
gitdb==4.0.10; python_version >= '3.7'
gitpython==3.1.40; python_version >= '3.7'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.4; python_version >= '3.5'
importlib-metadata==6.8.0; python_version < '3.10'
importlib-resources==5.12.0
iniconfig==2.0.0; python_version >= '3.7'
ipykernel==6.25.2
ipython==8.12.3; python_version >= '3.8'
ipython-genutils==0.2.0
ipywidgets==8.1.1; python_version >= '3.7'
isoduration==20.11.0
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
jupyter==1.0.0
jupyter-client==8.4.0; python_version >= '3.8'
jupyter-console==6.6.3; python_version >= '3.7'
jupyter-core==5.4.0; python_version >= '3.8'
jupyter-events==0.8.0; python_version >= '3.8'
jupyter-lsp==2.2.0; python_version >= '3.8'
jupyter-server==2.8.0; python_version >= '3.8'
jupyter-server-terminals==0.4.4; python_version >= '3.8'
jupyterlab==4.0.7; python_version >= '3.8'
jupyterlab-pygments==0.2.2; python_version >= '3.7'
jupyterlab-server==2.25.0; python_version >= '3.8'
jupyterlab-widgets==3.0.9; python_version >= '3.7'
kiwisolver==1.4.4
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.7.1
matplotlib-inline==0.1.6; python_version >= '3.5'
mistune==3.0.2; python_version >= '3.7'
nbclient==0.8.0; python_version >= '3.8'
nbconvert==7.9.2; python_version >= '3.8'
nbformat==5.9.2; python_version >= '3.8'
nest-asyncio==1.5.8; python_version >= '3.5'
notebook==7.0.6; python_version >= '3.8'
notebook-shim==0.2.3; python_version >= '3.7'
numpy==1.24.3
overrides==7.4.0; python_version >= '3.6'
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.5.0
pkgutil-resolve-name==1.3.10; python_version < '3.9'
plac==1.4.0
platformdirs==3.11.0; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
prometheus-client==0.17.1; python_version >= '3.6'
prompt-toolkit==3.0.39; python_version >= '3.7'
psutil==5.9.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0
pulp==2.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.16.1; python_version >= '3.7'
pyparsing==3.0.9
pytest==7.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7; python_version >= '3.6'
pytz==2023.3
pyyaml==6.0.1; python_version >= '3.6'
pyzmq==25.1.1; python_version >= '3.6'
qtconsole==5.4.4; python_version >= '3.7'
qtpy==2.4.0; python_version >= '3.7'
referencing==0.30.2; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
reretry==0.11.8; python_version >= '3.7'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.10.6; python_version >= '3.8'
scipy==1.10.1
send2trash==1.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0
smart-open==6.4.0; python_version >= '3.6' and python_version < '4.0'
smmap==5.0.1; python_version >= '3.7'
snakemake==7.32.4
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
stack-data==0.6.3
stopit==1.1.2
tabulate==0.9.0; python_version >= '3.7'
terminado==0.17.1; python_version >= '3.7'
throttler==1.2.2
tinycss2==1.2.1; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
toposort==1.10
tornado==6.3.3; python_version >= '3.8'
traitlets==5.11.2; python_version >= '3.8'
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0; python_version < '3.10'
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7; python_version >= '3.7'
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4; python_version >= '3.8'
widgetsnbextension==4.0.9; python_version >= '3.7'
wrapt==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yte==1.5.1; python_version >= '3.7'
zipp==3.15.0
